+++ src/risk_measures/AVaR.jl
@@ -4,6 +4,8 @@
 #  file, You can obtain one at http://mozilla.org/MPL/2.0/.
 #############################################################################
 
+export AVaR
+
 """
     AVaR(beta::Float64)